broke WeaposnCard Tests down into smaller units
