+++ src/components/WeaponsCard/WeaponsCard.spec.js
@@ -124,171 +124,4 @@ describe('The Weapons Card', () => {
       expect(wrapper.text()).not.toContain('ROF');
     });
   });
-  describe('modifying weapons', () => {
-    let wrapper;
-    const modifyPanel = () => wrapper.find('.modifyWeaponPanel');
-    beforeEach(() => {
-      wrapper = mountAppWithStore(storeWithCreateCharacterView());
-      wrapper.find('#addFirearm').simulate('click');
-      gunList(wrapper).find('#M16').simulate('click');
-      selectedWeapons(wrapper).find('#modifyM16').simulate('click');
-    });
-    it('should be possible to change default magazine type when magazine options are available', () => {
-      modifyPanel().find('#M16MagAtIndex1').simulate('click');
-      expect(modifyPanel().find('.modifyMagazines').childAt(1).text()).toContain('30');
-    });
-    it('should update weight when primary mags changed', () => {
-      modifyPanel().find('#M16MagAtIndex1').simulate('click');
-      expect(wrapper.find('#WeaponStatAW').text()).toContain('1');
-      expect(wrapper.find('#WeaponStatWeight').text()).toContain('9');
-    });
-    it('should be possible to add custom magazine', () => {
-      modifyPanel().find('#addCustomMagazine').simulate('click');
-      wrapper.find('#customMagCapacityInput').simulate('change', {
-        target: { value: '18' },
-      });
-      wrapper.find('#customMagWeightInput').simulate('change', {
-        target: { value: '.65' },
-      });
-      wrapper.find('#customMagTypeInput').simulate('change', {
-        target: { value: 'Mag' },
-      });
-      wrapper.find('#submitCustomMag').simulate('click');
-      expect(modifyPanel().find('.modifyMagazines').childAt(3).text()).toContain('18');
-      expect(modifyPanel().find('.modifyMagazines').childAt(3).text()).toContain('0.65');
-      expect(modifyPanel().find('.modifyMagazines').childAt(3).text()).toContain('Mag');
-    });
-    describe('custom magazine gaurd clases', () => {
-      const enterMagCapacity = () => {
-        wrapper.find('#customMagCapacityInput').simulate('change', {
-          target: { value: '18' },
-        });
-      };
-      const enterMagType = () => {
-        wrapper.find('#customMagTypeInput').simulate('change', {
-          target: { value: 'Mag' },
-        });
-      };
-      const enterMagWeight = () => {
-        wrapper.find('#customMagWeightInput').simulate('change', {
-          target: { value: '.65' },
-        });
-      };
-      const enterMagWeightAndType = () => {
-        enterMagWeight(wrapper);
-        enterMagType(wrapper);
-      };
-
-      beforeEach(() => {
-        wrapper = mountAppWithStore(storeWithCreateCharacterView());
-        wrapper.find('#addFirearm').simulate('click');
-        gunList(wrapper).find('#M16').simulate('click');
-        selectedWeapons(wrapper).find('#modifyM16').simulate('click');
-        modifyPanel().find('#addCustomMagazine').simulate('click');
-      });
-      it('should only accept numbers for capacity', () => {
-        wrapper.find('#customMagCapacityInput').simulate('change', {
-          target: { value: 'six' },
-        });
-        enterMagWeightAndType(wrapper);
-        wrapper.find('#submitCustomMag').simulate('click');
-        expect(wrapper.text()).toContain('Please Enter Valid Data');
-      });
-      it('should only accept whole numbers for capacity', () => {
-        wrapper.find('#customMagCapacityInput').simulate('change', {
-          target: { value: '18.5' },
-        });
-        enterMagWeightAndType(wrapper);
-        wrapper.find('#submitCustomMag').simulate('click');
-        expect(wrapper.text()).toContain('Please Enter Valid Data');
-      });
-      it('should only accept numbers for weight', () => {
-        enterMagCapacity(wrapper);
-        wrapper.find('#customMagWeightInput').simulate('change', {
-          target: { value: 'point six five' },
-        });
-        enterMagType(wrapper);
-        wrapper.find('#submitCustomMag').simulate('click');
-        expect(wrapper.text()).toContain('Please Enter Valid Data');
-      });
-      it('should have a value at least two characters long for type', () => {
-        enterMagCapacity(wrapper);
-        enterMagWeight(wrapper);
-        wrapper.find('#customMagTypeInput').simulate('change', {
-          target: { value: '!' },
-        });
-        wrapper.find('#submitCustomMag').simulate('click');
-        expect(wrapper.text()).toContain('Please Enter Valid Data');
-      });
-    });
-    it('should be possible to modify firearm weight', () => {
-      modifyPanel().find('#modifyWeaponWeight').simulate('click');
-      wrapper.find('#modifyWeightNoteInput').simulate('change', {
-        target: { value: 'added torch' },
-      });
-      wrapper.find('#modifyWeightValueInput').simulate('change', {
-        target: { value: '.5' },
-      });
-      wrapper.find('#submitModifiedWeight').simulate('click');
-      expect(wrapper.find('#WeaponStatWeight').text()).toContain('9.2');
-    });
-    it('should render the weight modifications', () => {
-      modifyPanel().find('#modifyWeaponWeight').simulate('click');
-      wrapper.find('#modifyWeightNoteInput').simulate('change', {
-        target: { value: 'added torch' },
-      });
-      wrapper.find('#modifyWeightValueInput').simulate('change', {
-        target: { value: '.5' },
-      });
-      wrapper.find('#submitModifiedWeight').simulate('click');
-      expect(modifyPanel().text()).toContain('added torch');
-      expect(modifyPanel().text()).toContain('0.5');
-    });
-    it('should be possible to remove weapon weight modifications', () => {
-      modifyPanel().find('#modifyWeaponWeight').simulate('click');
-      wrapper.find('#modifyWeightNoteInput').simulate('change', {
-        target: { value: 'added torch' },
-      });
-      wrapper.find('#modifyWeightValueInput').simulate('change', {
-        target: { value: '.5' },
-      });
-      wrapper.find('#submitModifiedWeight').simulate('click');
-      wrapper.find('.removeModification').simulate('click');
-      expect(wrapper.find('#WeaponStatWeight').text()).toContain('8.7');
-      expect(wrapper.text()).not.toContain('added torch');
-    });
-    describe('modify weapon weight gaurd clauses', () => {
-      beforeEach(() => {
-        wrapper = mountAppWithStore(storeWithCreateCharacterView());
-        wrapper.find('#addFirearm').simulate('click');
-        gunList(wrapper).find('#M16').simulate('click');
-        selectedWeapons(wrapper).find('#modifyM16').simulate('click');
-        modifyPanel().find('#modifyWeaponWeight').simulate('click');
-      });
-      it('should only accept numbers for weight value', () => {
-        wrapper.find('#modifyWeightNoteInput').simulate('change', {
-          target: { value: 'added torch' },
-        });
-        wrapper.find('#modifyWeightValueInput').simulate('change', {
-          target: { value: 'one pound' },
-        });
-        wrapper.find('#submitModifiedWeight').simulate('click');
-        expect(wrapper.text()).toContain('Please Enter Valid Data');
-      });
-      it('should have a value entered in type nore feild', () => {
-        wrapper.find('#modifyWeightValueInput').simulate('change', {
-          target: { value: '1' },
-        });
-        wrapper.find('#submitModifiedWeight').simulate('click');
-        expect(wrapper.text()).toContain('Please Enter Valid Data');
-      });
-    });
-    it('should be possible to remove all mods', () => {
-      wrapper = mountAppWithStore(storeWithCreateCharacterView(testM1911A1WithMods()));
-      selectedWeapons(wrapper).find('#modifyM1911A1').simulate('click');
-      modifyPanel().find('.removeAllMods').simulate('click');
-      expect(modifyPanel().text()).not.toContain('TestMag');
-      expect(wrapper.find('#WeaponStatWeight').text()).toContain('3');
-    });
-  });
 });
